+++ src/main.rs
@@ -78,6 +78,19 @@ fn main() -> io::Result<()> {
     println!("You're conversing with {} model", &config.model);
     let mut ollama_client = OllamaClient::new(config.model.clone(), config.system_prompt.clone());
 
+    match ollama_client.verify() {
+        Ok(s) => println!("{}", s),
+        Err(e) => {
+            println!("\n\nModel is not available: {}", e);
+            println!(
+                "Check that Ollama is installed or run `ollama pull {}` to pull the model.",
+                config.model
+            );
+
+            std::process::exit(1);
+        }
+    }
+
     // Main conversation loop
     loop {
         // Prompt the user for input
@@ -138,6 +151,8 @@ fn main() -> io::Result<()> {
             &user_prompt,
             input_file_content.as_deref(),
         )?;
+
+        println!("{}", ollama_response);
 
         history.append_user_input(&user_prompt)?;